Rename to_text_paraphrase to to_mnemonic and change parameter
The new function takes the address of the instruction itself instead of the
address of the next instruction. Because this is a backward incompatible
change and the type of the parameter does not change, we rename the function
to force a compile time error for users and avoid silent breakage.